Phase 2: move states/observables/entropy/fermions/qec into dense_evolution.physics/
- git mv all 5 modules into dense_evolution/physics/
- new physics/__init__.py re-exports the public API (ghz_state,
  pauli_expectation/pauli_sum_expectation/pauli_hamiltonian_to_matrix,
  partial_trace/von_neumann_entropy/mutual_information,
  majorana_pauli_terms, pauli_commutes/compute_syndrome/erasure_aware_decode)
- backward-compat shims at all 5 old top-level paths (dense_evolution.states,
  .observables, .entropy, .fermions, .qec) -- unlike interop/, the subpackage
  name here differs from any single old module name, so the dotted import
  path DOES change and a shim is genuinely required this time
- fixed states.py's one relative import broken by the move
  (.topology -> ..topology, since topology.py hasn't moved into circuits/ yet)
- updated dense_evolution/__init__.py, the 5 docs/api/*.md mkdocstrings
  directives, and pyproject.toml packages list

Full suite: 796 passed, 17 skipped, 4 failed on first run -- all 4 failures
were MemoryPressureError (known resource-contention pattern on this
machine from repeated full-suite runs today), confirmed via isolated
re-run: all 4 pass individually.

Part of #76.
