+++ dense_evolution/__init__.py
@@ -20,8 +20,8 @@ from .mitigation import (richardson_extrapolate, zero_noise_extrapolation, polyn
                           richardson_extrapolate_jit, zero_noise_extrapolation_jit,
                           polynomial_extrapolate_jit, uhlmann_fidelity_jit, zne_density_matrix_jit)
 from .topology import entangling_layer
-from .observables import pauli_expectation, pauli_sum_expectation, pauli_hamiltonian_to_matrix
-from .states import ghz_state
+from .physics.observables import pauli_expectation, pauli_sum_expectation, pauli_hamiltonian_to_matrix
+from .physics.states import ghz_state
 from .utils.measurement import sample_counts, statevector_fidelity
 from .qft import qft
 from .random_circuit import random_circuit
@@ -30,9 +30,9 @@ from .harrison_tb import (ELEMENTS as HARRISON_ELEMENTS, ETA as HARRISON_ETA,
                            sp3_dimer_hamiltonian, zincblende_hamiltonian)
 from .vhd_tb import (MATERIALS as VHD_MATERIALS, sp3s_star_hamiltonian,
                       direct_gap_at_gamma, band_extrema_along_path)
-from .fermions import majorana_pauli_terms
-from .entropy import partial_trace, von_neumann_entropy, mutual_information
+from .physics.fermions import majorana_pauli_terms
+from .physics.entropy import partial_trace, von_neumann_entropy, mutual_information
 from .trotter import pauli_rotation_ops, trotter_evolve_ops
-from .qec import pauli_commutes, compute_syndrome, erasure_aware_decode
+from .physics.qec import pauli_commutes, compute_syndrome, erasure_aware_decode
 
 __version__ = "8.1.60"
